requirements.txt
aiosmtplib==2.0.2
annotated-types==0.5.0
anyio==3.7.0
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.2
fastapi==0.99.1
fastapi-mail==1.3.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==1.10.10
pydantic_core==2.0.1
python-decouple==3.8
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
